Separate the App component from index.jsx file
